flytt felles validering inn i tilhørende plass
